Refactor:forwardRef를 이용하여 큐레이션 상세 페이지에서 단 하나의 topBar 만 사용할 수 있도록 수정
