+++ app/feature/curation/components/CurationDetail/CurationDetailCardList.tsx
@@ -11,31 +11,22 @@ import { CurationDetailResponse } from "@feature/curation/queries/dto/curation-d
 import { twMerge } from "tailwind-merge";
 import { MyCurationResponse } from "@feature/curation/queries/dto/my-curation";
 import MapIcon from "@common/assets/icons/map/map";
-import { useEffect, useState } from "react";
+import { forwardRef, useEffect, useState } from "react";
 import Map from "@feature/map/components/Map/Map";
 
 interface CurationDetailCardListProps {
+  inView: boolean;
   curationId: number;
   curationDetail: CurationDetailResponse;
   myCurationData: MyCurationResponse;
 }
 
-export default function CurationDetailCardList({
-  curationId,
-  curationDetail,
-  myCurationData,
-}: CurationDetailCardListProps) {
-  const {
-    cardRefs,
-    openScrapDeleteToast,
-    toastText,
-    placeIndex,
-    scrollHeight,
-    handlers,
-  } = useCurationDetailCardList(curationDetail.spaceDetails);
-  const [inViewRef, inView] = useInView({
-    threshold: 1,
-  });
+const CurationDetailCardList = forwardRef<
+  HTMLDivElement,
+  CurationDetailCardListProps
+>(({ ...props }, ref) => {
+  const { cardRefs, openScrapDeleteToast, toastText, placeIndex, handlers } =
+    useCurationDetailCardList(props.curationDetail.spaceDetails);
   const [mapOpen, setMapOpen] = useState(false);
   const [mapPlaceData, setMapPlaceData] = useState<
     {
@@ -57,7 +48,7 @@ export default function CurationDetailCardList({
 
   useEffect(() => {
     setMapPlaceData(
-      curationDetail.spaceDetails.map((space) => {
+      props.curationDetail.spaceDetails.map((space) => {
         return {
           address: space.address,
           name: space.name,
@@ -67,20 +58,13 @@ export default function CurationDetailCardList({
         };
       })
     );
-  }, [curationDetail]);
+  }, [props.curationDetail]);
   return (
     <>
-      {!inView && (
-        <div className="w-[100%] absolute top-0 z-10">
-          <CurationTopAppBar
-            curationId={curationId}
-            curationDetail={curationDetail}
-            text={curationDetail.title}
-            variant={curationDetail.variant}
-            className="bg-white"
-          />
+      {!props.inView && (
+        <div className="w-[100%] absolute top-[4.8rem] z-10">
           <div className="flex bg-white z-10 gap-[0.8rem] pb-[0.8rem] pt-[0.6rem] pl-[2rem] overflow-x-scroll">
-            {curationDetail.spaceDetails.map((item, i) => (
+            {props.curationDetail.spaceDetails.map((item, i) => (
               <Filter
                 key={i}
                 photo={item.imageUrls && item.imageUrls[0]}
@@ -88,7 +72,8 @@ export default function CurationDetailCardList({
                 selected={placeIndex === i}
                 className={twMerge(
                   "whitespace-nowrap",
-                  curationDetail.spaceDetails.length - 1 === i && "mr-[1.2rem]"
+                  props.curationDetail.spaceDetails.length - 1 === i &&
+                    "mr-[1.2rem]"
                 )}
                 onClick={() => handlePlaceFilterClick(i)}
               />
@@ -102,7 +87,7 @@ export default function CurationDetailCardList({
             <div className="flex items-center gap-[0.4rem]">
               <LocationFillIcon />
               <h3 className="text-black body2-medium">
-                {curationDetail.spaceDetails.length}개의 공간
+                {props.curationDetail.spaceDetails.length}개의 공간
               </h3>
             </div>
             <button
@@ -114,10 +99,10 @@ export default function CurationDetailCardList({
             </button>
           </div>
           <div
-            ref={inViewRef}
+            ref={ref}
             className="flex gap-[0.8rem] mb-[-10.6rem] overflow-x-scroll"
           >
-            {curationDetail.spaceDetails.map((item, i) => (
+            {props.curationDetail.spaceDetails.map((item, i) => (
               <Filter
                 key={i}
                 photo={item.imageUrls && item.imageUrls[0]}
@@ -125,7 +110,8 @@ export default function CurationDetailCardList({
                 selected={placeIndex === i}
                 className={twMerge(
                   "whitespace-nowrap",
-                  curationDetail.spaceDetails.length - 1 === i && "mr-[1.2rem]"
+                  props.curationDetail.spaceDetails.length - 1 === i &&
+                    "mr-[1.2rem]"
                 )}
                 onClick={() => handlePlaceFilterClick(i)}
                 // onClick 하면 get Data 변경
@@ -133,34 +119,19 @@ export default function CurationDetailCardList({
             ))}
           </div>
         </div>
-        {curationDetail.spaceDetails.map((props, i) => (
+        {props.curationDetail.spaceDetails.map((space, i) => (
           <CurationDetailInfoCard
-            key={props.name}
-            variant={curationDetail.variant}
-            curationId={curationId}
-            {...props}
+            key={space.name}
+            variant={props.curationDetail.variant}
+            curationId={props.curationId}
+            {...space}
             ref={cardRefs[i]}
             handleDeleteToast={handlers.changeOpenScrapDeleteToast}
             handleToastText={handlers.changeToastText}
-            myCurationData={myCurationData}
+            myCurationData={props.myCurationData}
           />
         ))}
       </div>
-      {scrollHeight > 370 && (
-        <div className="flex gap-[0.8rem] w-full bg-white pl-[2rem] py-[0.4rem] pb-[0.8rem] overflow-x-scroll fixed top-[5.4rem]">
-          {curationDetail.spaceDetails.map((item, index) => (
-            <Filter
-              key={index}
-              photo={item.imageUrls ? item.imageUrls[0] : undefined}
-              label={item.name}
-              selected={placeIndex === index}
-              className="whitespace-nowrap"
-              onClick={() => handlePlaceFilterClick(index)}
-              // onClick 하면 get Data 변경
-            />
-          ))}
-        </div>
-      )}
       {mapOpen && (
         <Map
           placeData={mapPlaceData}
@@ -172,4 +143,8 @@ export default function CurationDetailCardList({
       <Toast open={openScrapDeleteToast} text={toastText} />
     </>
   );
-}
+});
+
+CurationDetailCardList.displayName = "CurationDetailCardList";
+
+export default CurationDetailCardList;
